test: LineTest 중첩 구조를 통한 가독성 개선
